[movies] analysis: Disable dialog fragment callbacks detector lint rule
This 'DialogFragmentCallbacksDetector' rule is making Lint crash with a
generic Lint error and as such is disabled until it gets fixed in a
newer version of the fragment library or agp.

Also, as part of this, the 'UnknownIssueId' is disabled as well as it
seems that the 'DialogFragmentCallbacksDetector' rule is not recognized.