[movies] analysis: Suppress monochrome launcher icon lint warning
Details: Monochrome icon is not defined

If android:roundIcon and android:icon are both in your manifest, you
must either remove the reference to android:roundIcon if it is not
needed; or, supply the monochrome icon in the drawable defined by the
android:roundIcon and android:icon attribute.

For example, if android:roundIcon and android:icon are both in the
manifest, a launcher might choose to use android:roundIcon over
android:icon to display the adaptive app icon. Therefore, your themed
application icon will not show if your monochrome attribute is not also
specified in android:roundIcon.