[movies] analysis: Enable running detekt with type resolution
This change fixes all type resolution specific warnings and then enables
running Detekt with type resolution.

As such, with this change, and in addition to running Detekt via
'./gradlew detekt' task, which runs Detekt without type resolution,
Detekt is now able to run with type resolution on the 'main' and 'test'
source sets. This is possible via running the below tasks:
- Main Source: './gradlew detektMain'
- Test Source: './gradlew detektTest'