Deps: Disable android gradle plugin version lint error
Error Message: "A newer version of com.android.tools.build:gradle than 8
.1.0-beta01 is available: 8.2.0-alpha02 [AndroidGradlePluginVersion]"

Explanation: "This detector looks for usage of the Android Gradle Plugin
where the version you are using is not the current stable release. Using
older versions is fine, and there are cases where you deliberately want
to stick with an older version. However, you may simply not be aware
that a more recent version is available, and that is what this lint
check helps find."

------------------------------------------------------------------------

This 'AndroidGradlePluginVersion' Lint error is being disabled as it
causes problems with automated PRs, such as automated Renovate or
Dependabot dependency update PRs, which fail on CI due to Lint failing
overall, causing the PR checks to block those PRs.