Deps: Add package rules for the compose compiler, kotlin and ksp updates
The 'Compose Compiler' has a strict mapping to specific versions of
Kotlin Typically, these two dependencies aren’t released at the same
time, so there is a delay of several days between Kotlin’s new version
and the corresponding 'Compose Compiler'. This means that you can’t
update Kotlin to the latest version if the 'Compose Compiler'
hasn’t been released yet. Same applies with 'KSP' and 'Kotlin'.

With this Renovate configuration change, the 'Compose Compiler',
'Kotlin' and 'KSP' updates will get grouped together in order to avoid
having multiple PRs opened (3), which will be anyway unable to get
merged as is and in isolation to each other. Instead, with that change,
one PR will be opened, which will be able to get merged as those updates
will be now grouped in that single PR instead.

For more info see: Automating dependency updates in a Compose project
- https://medium.com/androiddevelopers/automating-dependency-updates-in-
a-compose-project-168ef5e89ac5