Build: Add extra on unused dependencies configuration for java io file
This is based on build health status, which started reporting these
unused dependencies, but for no particular reason. Maybe this is related
to the previous 'Compose Compiler', 'Kotlin' and 'KSP' update, or the
extra KSP related 'jvmToolchain' configuration, or even the fact that
the whole 'datastore' module got temporarily removed.